Refactor logging service and common module
