+++ src/common/common.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { LoggingService } from './common.logging.service';
+
+@Module({
+  providers: [LoggingService],
+  exports: [LoggingService],
+})
+export class CommonModule {}
