+++ src/common/logger.middleware.ts
@@ -1,6 +1,6 @@
 import { Injectable, NestMiddleware } from '@nestjs/common';
 import { Request, Response, NextFunction } from 'express';
-import { LoggingService } from './logging.service';
+import { LoggingService } from './common.logging.service';
 
 @Injectable()
 export class LoggerMiddleware implements NestMiddleware {
